service: transfer TTL code to a separate file
