+++ service/meta.go
@@ -1,11 +1,5 @@
 package service
 
-import (
-	"github.com/pkg/errors"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-)
-
 type (
 	// MetaHeader contains meta information of request.
 	// It provides methods to get or set meta information meta header.
@@ -15,9 +9,8 @@ type (
 		ResetMeta() RequestMetaHeader
 		RestoreMeta(RequestMetaHeader)
 
-		// TTLRequest to verify and update ttl requests.
-		GetTTL() uint32
-		SetTTL(uint32)
+		// TTLHeader allows to get and set TTL value of request.
+		TTLHeader
 
 		// EpochHeader gives possibility to get or set epoch in RPC Requests.
 		EpochHeader
@@ -46,20 +39,6 @@ type (
 		GetRaw() bool
 		SetRaw(bool)
 	}
-
-	// TTLCondition is closure, that allows to validate request with ttl.
-	TTLCondition func(ttl uint32) error
-)
-
-const (
-	// ZeroTTL is empty ttl, should produce ErrZeroTTL.
-	ZeroTTL = iota
-
-	// NonForwardingTTL is a ttl that allows direct connections only.
-	NonForwardingTTL
-
-	// SingleForwardingTTL is a ttl that allows connections through another node.
-	SingleForwardingTTL
 )
 
 // SetVersion sets protocol version to ResponseMetaHeader.
@@ -70,9 +49,6 @@ func (m *ResponseMetaHeader) SetEpoch(v uint64) { m.Epoch = v }
 
 // SetVersion sets protocol version to RequestMetaHeader.
 func (m *RequestMetaHeader) SetVersion(v uint32) { m.Version = v }
-
-// SetTTL sets TTL to RequestMetaHeader.
-func (m *RequestMetaHeader) SetTTL(v uint32) { m.TTL = v }
 
 // SetEpoch sets Epoch to RequestMetaHeader.
 func (m *RequestMetaHeader) SetEpoch(v uint64) { m.Epoch = v }
@@ -91,42 +67,3 @@ func (m *RequestMetaHeader) ResetMeta() RequestMetaHeader {
 
 // RestoreMeta sets current RequestMetaHeader to passed value.
 func (m *RequestMetaHeader) RestoreMeta(v RequestMetaHeader) { *m = v }
-
-// IRNonForwarding condition that allows NonForwardingTTL only for IR
-func IRNonForwarding(role NodeRole) TTLCondition {
-	return func(ttl uint32) error {
-		if ttl == NonForwardingTTL && role != InnerRingNode {
-			return ErrInvalidTTL
-		}
-
-		return nil
-	}
-}
-
-// ProcessRequestTTL validates and update ttl requests.
-func ProcessRequestTTL(req MetaHeader, cond ...TTLCondition) error {
-	ttl := req.GetTTL()
-
-	if ttl == ZeroTTL {
-		return status.New(codes.InvalidArgument, ErrInvalidTTL.Error()).Err()
-	}
-
-	for i := range cond {
-		if cond[i] == nil {
-			continue
-		}
-
-		// check specific condition:
-		if err := cond[i](ttl); err != nil {
-			if st, ok := status.FromError(errors.Cause(err)); ok {
-				return st.Err()
-			}
-
-			return status.New(codes.InvalidArgument, err.Error()).Err()
-		}
-	}
-
-	req.SetTTL(ttl - 1)
-
-	return nil
-}
